refactor: consolidate common headers
Merge calibrate.h + benchmark_constants.h into common.h.
Absorb BackendGating into Peak base class.
Remove Timer class, inline std::chrono at call sites.
Extract OpenCL-only utilities (populate, roundToMultipleOf,
trimString, UNUSED) to opencl/cl_utils.h/.cpp.
Fix result_store.cpp OS_NAME duplication.

Net: 10→8 headers, 9→8 sources, backend includes 7→4.
