+++ src/common/AGENTS.md
@@ -1,8 +1,8 @@
 # src/common — Shared Backend-Neutral Code
 
-Base classes, utilities, result store, calibration, and inventory shared
+Base classes, utilities, result store, and inventory shared
 by every backend. Does NOT contain a logger implementation — each consumer
-(desktop CLI, Android) provides its own (`src/cli/logger.cpp` or
+(desktop CLI, Android) provides its own (`src/cli/logger_cli.cpp` or
 `android/.../logger_android.cpp`).
 
 ## Quick Lookups
@@ -10,8 +10,8 @@ by every backend. Does NOT contain a logger implementation — each consumer
 - Looking for the Peak base class? → `peak.cpp` + `include/common/peak.h`
 - Understanding result recording API? → `include/common/logger.h` (header only)
 - Understanding result output format? → `result_store.cpp` + `include/common/result_store.h`
-- Understanding CLI option gating? → `backend_gating.cpp` + `include/common/backend_gating.h`
-- Understanding calibration? → `calibrate.cpp` + `include/common/calibrate.h`
+- Understanding calibration? → `common.cpp` (`pickIters()`) + `include/common/common.h`
+- Understanding gating? → `peak.cpp` + `include/common/peak.h` (gating lives in Peak)
 - Adding a new backend? → The `Peak` interface is in `include/common/peak.h`
 - Understanding device inventory structs / JSON? → `inventory.cpp` + `include/common/inventory.h`
 
@@ -19,12 +19,10 @@ by every backend. Does NOT contain a logger implementation — each consumer
 
 | File | Purpose |
 |------|---------|
-| `peak.cpp` | `Peak` base class: constructor, `applyOptions()` common logic |
-| `common.cpp` | `benchmark_config_t::forDevice()`, `Timer`, `populate()`, utilities |
+| `peak.cpp` | `Peak` base class: `applyOptions()` copies CLI state (including gating) |
+| `common.cpp` | `benchmark_config_t::forDevice()`, `pickIters()` calibration |
 | `result_store.cpp` | `ResultEntry`/`ResultStore` serialization: JSON, CSV, XML |
 | `logger.cpp` | Base `logger` class: result-scope API, `emit()`, `recordSkip` |
-| `backend_gating.cpp` | `BackendGating::copyFrom()` — copies CLI gating state |
-| `calibrate.cpp` | `pickIters()` — runtime iteration calibration |
 | `inventory.cpp` | `inventoryToJson()` — device inventory JSON serializer (no backend includes) |
 | `options.cpp` | `parseCliOptions()` — CLI argument parsing |
