--- src/common/backend_gating.cpp
@@ -1,7 +1,0 @@
-#include <common/options.h>
-#include <common/backend_gating.h>
-
-void BackendGating::copyFrom(const CliOptions &opts) {
-  enabledTests      = opts.enabledTests;
-  enabledCategories = opts.enabledCategories;
-}
