+++ src/common/peak.cpp
@@ -1,13 +1,12 @@
 #include <common/peak.h>
 #include <common/options.h>
 
-Peak::Peak() = default;
-
 void Peak::applyOptions(const CliOptions &opts)
 {
     forceIters     = opts.forceIters;
     specifiedIters = opts.iters;
     warmupCount    = opts.warmupCount;
     targetTimeUs   = opts.targetTimeUs;
-    gating.copyFrom(opts);
+    enabledTests      = opts.enabledTests;
+    enabledCategories = opts.enabledCategories;
 }
