+++ src/opencl/CMakeLists.txt
@@ -58,6 +58,7 @@ endif()
 add_library(peak_opencl STATIC
     cl_peak.cpp
     cl_common.cpp
+    cl_utils.cpp
     global_bandwidth.cpp
     local_bandwidth.cpp
     image_bandwidth.cpp
