+++ src/opencl/local_bandwidth.cpp
@@ -5,7 +5,7 @@ int clPeak::runLocalBandwidthTest(cl::CommandQueue &queue, cl::Program &prog, de
   float timed, gbps;
   cl::NDRange globalSize, localSize;
 
-  if (!gating.isAllowed(Benchmark::LocalBW))
+  if (!isAllowed(Benchmark::LocalBW))
     return 0;
 
   unsigned int forced = forceIters ? specifiedIters : 0;
